README.md
# IMDb-Movies-and-TV-Shows
This project focuses on analyzing and predicting the success of movies and TV shows based on data from IMDb (Internet Movie Database). The goal is to build a predictive model that can estimate the rating, popularity, or success of a movie/TV show using features such as genre, cast, director, runtime, release year, and more.